Bugfix: treat NaN as zero and let zero divided by zero equal zero
Invalid values were produced in the twopoint ray traicng algorithm because the authors rely on "guards" to mask these errors. If a division would produce an invalid value, for example because the denominator is zero or the radicand is below zero, the guard clause will be zero for this case. Multiplying by this guard clause effectively removes the invalid value. Python (or numpy) treats multiplying a NaN by zero as another NaN, so these values are never caught. For now, treat NaNs as zero and maybe look for a better way