Replace xtensor in two point ray tracing by valarray
First, the xtensor implementation had a bug which could be fixed by explicitly replacing xminus2 by a double instead of keeping it as an unevaluated 0D xtensor. This fix seemed dirty and I didn't know where similar bugs would crop up from the lazy nature of xtensor where a temporary reference was evaluated after free. As a side effect of the rewrite, the code is 3x faster (11µs vs 34µs) for a representative test case.