Rewrite unit vector function
Old code used slowness vector and the connection vector between start and end to define a plane. If these two were almost the same direction, it would fall back to define a vector perpendicular to that direction. This vector would not have to be in the rays plane, meaning the function didn't work as intended.