Bugfix: Distance between end and found segment was taken as index
Finding the correct index of a segment containing a given arclength was done by walking backwards until the first segment with begin arclength <= given arclength was found. But instead of the distance between the begin and the found segment distance to the end was calculated. This resulted in the wrong index being returned, since indexing starts at 0 at the begin of the beam.