Fix build error in unittests due to missing header
doublebeam_lib has fmt in its headers, for example in kdtree hpp, so the dependencies which include those headers (even transitively) need to have the include path or the header has to be in the system include paths.
Make the link dependency public means all users will have fmt available, even when not installed system wide.